Don't give a hard error if we can't determine the fingerprint for a custom toolchain
Previously, cargo sweep would unconditionally give an error when it couldn't determine the
fingerprint. But for custom toolchains, it makes sense for it to be linked but give an error when
you try to run it: it means the compiler hasn't yet been built.

Rather than aborting straight away, treat the toolchain as if it isn't installed and clean its artifacts.