Fix the empty_project_output test for macos
The empty_project_output test doesn't have the same traversing/logging
order on macos as it does on whatever os/machine it was written for.
This commit tries to improve the situation somewhat by making the test
runnable on macos as well, while sacrificing readability and test accuracy.

An alternative fix would be to not use regex for the test at all.